Add a local FastBoot setup
This makes it easier to test the dummy app against a specific environment without having to deal with local untracked code changes.

A .env file can be created (starting from the .env.example file) to configure the needed env variables.

The `start:env` script can then be used to start the test app.